router: refactor profile relationships
In an effort to make testing and a no-alloc build much easier this
commit completely detaches profiles from the router or any transmission
object. By detching them a no alloc build requires no changes to the
profiles as they are now just wrapped to be installed. This also removes
all weak references as well and the Rc<RefCell>> on the router.

The downside is this adds a bit of complexity in the transmission path
1. it requires profiles only to transmit when asked if they have
   anything. This means all state must be queued up and sent when tx
   time is requested. This complicates profiles but eases the relation
   graph immensily and also is significantly more testable.
2. In an effort to keep Tx types as similar to trait objects as possible
   but not actually depending on core, the TxMessage enum is added. Its
   just a enum of all Tx types like RxMessageType but it implements the
   AntTxMessageType directly so it can be treated in a way like a trait
   object.