ant+: helpers: move entirely to callback model
Anything other than callbacks requires caching or state management and
its a pain or uses more memory than needed. Due to the nature of config
and data having two different implications when being said, use helper
types to limit what is being requested so we don't accidentally get
something we are not ready to handle.