Initial rework to mpsc channels
Start of breaking the connection between channels and the router. This
allows profiles to be a lot more flexible in their execution domain and
order. It also breaks a lot of borrowing limitations being impossed.

Future work will expose traits so any type of mpsc channel can be used.
For initial implemenation thingbuf will be used.